flow: handle struct initialization better
There problem with the original code was that if you had nested struct then
only the outer and first inner struct were initialized to zero.

In the original code, it tried to keep track of which values had been
initialized and which had not.  In the new code, it just looks it up in
SMATCH_EXTRA to see if it has been initialized.  This is simpler.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>