precedence: complain about "foo << bar - baz"
The negate operation has higher precedence then the shift so
"foo << bar - baz" can be re-written as "foo << (bar - baz)".  When you see
these in the code, it's normally a bug.  There are some macros like
Dbl_right_align() in parisc where people use "foo << bar - baz" and it's
not a bug, but that's ugly code so I don't feel bad complaining about it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>