type: make sval_type_max() default to "long long"
sval_type_min() defaults to "long long" so this makes it symmetric.
Ideally we wouldn't have places where we don't know the type but in real
life this can happen for two reasons.  First are bugs in Smatch.  Second
sometimes we just can't tell the type.

Defaulting to long long is ok because it marks everything as unknown.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>