user_data2: tag it if a function sets parameters to user data
This has always been tricky because in the end you need to re-implement
PARAM_SET/FILTER sort of...  I guess this patch sort of re-implements
PARAM_SET.  It says that if a returned parameter is always user data then
mark it like that.  But if it's not always user data then ignore it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>