modification_hooks:  set a variable as modified when it goes out of scope
This affects smatch_stored_condition.c because that stores the condition
without a sym pointer.  So if you have two "err" variables in different
scopes and we end the scope with "err == 0" then the next time we see
"if (err) " it complains that we already have checked this condition.

Now we clear all the stored "err" conditions at the end of "err's" scope.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>