function_hooks: add a hack around in compare_db_return_states_callbacks
The problem here is that when we set the state in a fake tree, we also
set the state in a real tree.  My feeling is that we shouldn't do that
but we rely on this behavior in many places.  I haven't figured out a
good way to fix this.

So for now, I'm just adding this very narrow hack around the problem.

In function hooks say we set the state on one path but not on the others
then it gets set on all the paths because we set the real cur_stree as
well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>