type_to_str: fix a crashing bug
We end up referencing freed memory and it segfaults.

I am unhappy with reverting this code but I also just want to fix it in a
hurry.  It's not totally clear how this even happens.  I feel like it
means there are other bugs in here.

I'm unhappy generally with the inconsistent way strings are allocated and
freed in smatch.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>