fixup_kernel: kmalloc() can allocate more data.  Let's say 4MB.
The actual amount of data that kmalloc() can allocate depends on the
config.  I'm not sure why I have 128k here.  I think I meant 128MB, since
that is a the answer in some configs.  There is at least one place which
relies on kmalloc() to allocate a 1MB chunk of data.  Smatch was assuming
it failed every time.  That creates hard to understand false positives.

So let's boost it up and say that kmalloc can allocate up to 4MB.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>