type: improve get_binop_type()
In the original code it returned NULL if we didn't know either the ->left
or ->right type.  But actually for some situations like "1 << foo" the type
of the binop is determined by the left hand side so we know the type even
if we don't know the type of the right hand side.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>