function_hooks: fix validation/sm_compare15.c handling
This is still not 100% perfect...  But anyway, if we have code like:

	if (foo < some_func()) {

I guess normally the assumption is that we know the return value of
some_func() and are checking the value of "foo".  But sometimes we know
"foo" and are checking what some_func() returns.  So we need to filter
both sides...

The part that isn't right is what about if both sides are a function call?

This stuff needs to be re-written.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>