assigned_expr: store it late
The problem here is say we have something like:

	struct whatever *foo = &some_struct->member_struct;


What happens is that first we say foo->bar and
some_struct->member_struct->bar are the same memory.  Then we say, "Oh,
this is a struct assignment, lets set foo->bar which ends up setting
some_struct->member_struct->bar."  Then if some_struct is a parameter we
record that the values got reset in this function, which is not true.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>