extra: handle x = (u8)y; correctly
If we do a truncate cast then that should be reflected when we handle
the assignment.  Currently we just strip the cast away.

When this was introduced, I said that I changed "how casts are handled" but
I think it is an "improvement".  Which is terribly vague and it's obviously
not correct to ignore casts...

Fixes: 1dcfc9fdad6d ('extra: shuffle match_assign a bit')
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>