Documentation: make me less confused
It is not immediately obvious that one is supposed to be in the kernel
source tree when running smatch_scripts/build_kernel_data.sh (I was
looking for an option to indicate that directory). The help text in
the script is also confusing, since that can easily be interpreted as
a statement about one's PWD, not the location of the script itself.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>