db: fix a bug in split returns
The situation here is that we have code like:

	return ret;

We are trying to figure out how the implications are if ret = -ENOMEM or
ret = 0.  The problem is that it would first find the implications of
-ENOMEM then it would take assume all those were true and find the
implications of = 0.  We should reset the implications back to the start
before we try find the next set of implications.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>