buf_size: improve resizable struct handling
If we have an zero size array on the end of a struct then Smatch uses the
stored size of the struct allocation to figure out the size of the array.
But if the last member of the struct is a struct with a zero size array
then we should be able to handle that as well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>