debug: add some tools for developing the db parser some more
I have a bunch of work in progress stuff.  I was working on changing how
the database parsed parameters and this was supposed to help me test it.

You give it a type and a recipe and it prints a value range.  See
validation/sm_val_parse1.c for an example.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>