user_data2: handle pointers to user data better
There are a couple related issues here.  If you had a buffer containing
user data then we should record that *buf is USER_DATA.  But even if it
is a struct pointer containing user data and we have tagged the members,
we should still record that *p is USER_DATA in case the buffer gets
casted to something else.

This information needs to be recorded in caller_info etc.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>