math: give up if calculating a value is too complicated
The kernel has a macro called BLK_RING_SIZE() which expands into a whole
wall of text when you precompile it.  The recursion count is over 10k.  It
leads to a compile failure because you run out of memory.

I have capped the recursion count at 200.  This seemed like a reasonable
guess.  When I tested it, there were about 40 kernel files affected.  This
seems reasonable.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>