extra: start tracking "chunks" instead of just variables
If you had a condition:

	if (foo > 42)

Then smatch would track the value of "foo" but if you had a condition:

	if (foo + bar > 42)

then we want smatch to track that as well.  I'm calling "foo + bar" a
chunk.  Much of this code is similar to smatch_stored_conditions.c.  You
have to create a var_sym_list of the links.  If any of them are modified
then it destroys the state of the chunk.

Of course, it's very primitive for now, and will be expanded in later
patches.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>