chunk: change the API a bit
There are two ways we can call expr_to_chunk.  First if we have a new chunk
and second if we are looking up an old value.  I have some ideas for how
both of them could work in the future but the main thing here is that I
wanted to tie the vsl in with the get chunk instead of doing it as a
separate call.  They are closely related actions.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>