chunk: improve expr_to_chunk_sym_vsl()
I the old system expr_to_chunk_sym_vsl() could handle "if (foo + bar)"
when "foo" and "bar" were simple variables.  Now it can handle a wide
range of stuff which involves up to two variables...

There are two sort of disadvantages, it ignores global variables since
those could be modified in a different thread.  Also in the old system
it reordered expressions so that it knew that "foo + bar" and "bar + foo"
were equivalent and now it doesn't do that.  Don't write messy code is
the rule.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>