pointer_math: don't complain if a pointer math operation is within bounds
I'm trying to make the pointer math much more verbose and one of the
common false positives that I'm seeing with my new code is when we are
copying data from a char array to an array with different size elements.
It maybe looks like this:

	for (i = 0; i < sizeof(char_array); i++)
		other_array[i] = char_array[i];

If we know that other_array[i] doesn't write out of bounds then don't
complain.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>