ranges: fix how some ranges are joined together
The common case here is that we added a range "0-u32max" and now we are
deleting all the other ranges in the list because they are included in
the big range.

But say we added one and it partially overlaps with the next for example,
0-5 and 4-10.  Then we should make 0-5 into 0-10 and delete the 4-10 range.
The old code just made it, "0-5,6-10" which isn't terrible but it's not
ideal.  (I'm not smart enough to write test cases for this, but this is
just from reading the code).

I changed == to return directly instead of looping through one more time
and then returning.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>