err_ptr: use the actual values to silence false positives
This check used to work like this:
1) Make a list of functions which return ERR_PTR().
2) If the caller doesn't check for IS_ERR() before dereferencing then
complain.

The problem is that some functions instead of checking IS_ERR() directly
called to helper functions that checked for IS_ERR() along with other
types of validation.  Because of cross function analysis, then
smatch_extra.c understands that the pointer is not an ERR_PTR() but this
check would still complain because it wasn't using the value.

Fixed now.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>