overflow: silence glibc strcmp false positives with string literal
The glibc strcmp() and strncmp() macros produce a lot of false positives
when you pass them short string literals.  If you read beyond the end of
a string literal, it's not really harmful.  They do this because it is
faster.  (I am going on what I remember from the last time I tried to
silence these warnings).

Reported-by: Karel Zak <kzak@redhat.com>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>