stored_conditions, implied: use implications of previous stored conditions
This only looks at binop conditions, because presumably smatch_extra.c and
smatch_comparison.c are handling the others fine.  In the future, I
imagine that we will try to use all three combined implications at the
same time, but I will have to test that a bit.  But basically this
parses code like:

	a = 1;
	if (foo & 12)
		a = 2;

	if (foo & 12)
		__smatch_implied(a);

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>