buf_size, db: only store type_size if it is a pointer to a SYM_BASETYPE
The function_type_size table has grown huge and most of the data is
total garbage.  Every time pointer was assigned it would say I don't
know how big of an array we are assigning to list->next, but the problem
is that list->next is just one struct and not an array at all.  But if
a struct member is a pointer to int or other SYM_BASETYPEs then that
probably is an array so let's save that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>