modification_hook: record the modification expression
One of the problems I have is that the released version of Smatch and the
version I use are out of sync.  It is a recurring headache.  Anyway, this
is something I have been thinking about for a while, but I don't have an
immediate user for...

One possible use would be to re-write the canonical for loop handling using
this.

It's sort of like get_assigned_expr() except it stores other sorts of
modification as well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>