double_checking: avoid more false positives
Oleg reported that there are false positives caused by backwards gotos.
Silence this by adding an "if (inside_loop()) " test.

If it is a known value then we don't need to check.

We seem not to need the code marked "FIXME:  needed?" so I have deleted it.
I guess we do use "is_true" for ignoring duplicate checks to see that a
variable is non-NULL.  I've just swapped that for an implied_condition_true()
test.

Reported-by: Oleg Drokin <green@linuxhacker.ru>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>