user_data2: tighten user data range lists
How user_data2 works is that if you have:

	foo = <user_controled> + bar;

Then is calculates the user controlled values of the right side and stores
that as the state of foo.  Then when a check needs to know the user
controlled range list of foo, it takes the state and it also takes the
state from smatch_extra.c and the intersection of both range lists is the
possible user controlled values for foo.

The problem is that you have something like:

	foo = user_controlled 1 - u32max;
	if (foo > 8) {
		foo = not user controlled 1 - u32max;
	} else {
		__smatch_user_rl(foo);
	}
	__smatch_user_rl(foo);

For the first __smatch_user_rl() the state of user_data2 is 1-u32max and
the state of smatch_extra is 1-8 so it prints the correct value.  But for
the next the result is that the state of user_data2 is 1-u32max and
smatch_extra.c is also 1-u32max.

We need to set user_data2 in a pre_merge_hook.  The odd thing about the
pre merge hook is that the sm is from the opposite side of the merge from
the cur_stree.  I needed that behavior for the first time I used this hook
but it's sort of unexpected...  I need to re-write that hook better, later.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>