equiv: mark variables as equiv even when their types are different
The old rule was that for variables to be equivalent then their types had
to be the same.  Now the rule is that the value inside the variable has
to fit into both variables for it to be equiv.

So say you have:  "x = y;" where x is an unsigned int and y is a signed
int.  If y can be negative then they are not equiv but if y is 0-4 then
they are equiv.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>