extra: fix get_user() handling
When we are parsing "ret = get_user();" then we end up with a fake
assignment which looks like: "ret = __builtin_expr(result, 0);"  The
problem is that we use strip_expr() in smatch_flow.c so it sees it as
"ret = result;" which is not a function call.  But then in smatch_extra.c
we don't use strip_expr() because we care about casting so we see it as
a function call.  In smatch extra we ignore function calls because
smatch_function_hooks.c handles those.  But because of the thing in
smatch_flow.c it is assumed that those are handled in smatch_extra.c.

In the end everyone ignores it.

Fix this by making removing the __builtin_expect() in smatch_extra.c so
it is not treated as a function call.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>