function_hooks: impossible: don't set both sides to impossible by mistake
The issue here is that we are trying to make sure a state is set on both
the true and false side.  Before we used to save the orig states to both
the true and false sides and then overwrite them.  But the bug is that say
there is another condition hook (the hook in smatch_impossible) which is
run before the function_hooks.  It sets the true path to &impossible and
that also sets the cur_stree it &impossible as well, and it also sets the
false path to &possible.  Now in the function_hooks we take the
"original states" from cur_stree (&impossible) and overwrite both sides,
and then re-write it with whatever function hook states there are.

What we should do is only overwrite both sides if another hook has not
already set the state.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>