flow: fix some issues with too hairy functions
We shouldn't be calling the END_FUNC_HOOK or AFTER_FUNC_HOOK when we do a
fake return.  That's not a part of a return statement.  Because we called
these more than once, it lead to some use after free issues.  In the kernel
this affected fs/btrfs/file.c parsing sometimes (For me it was when I had
--info enabled).

Also we were calling fake_a_return() for every statement instead of just
the first one.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>