user_data2: use fake assignments instead of duplicating that (badly)
Say you have a struct assignment "foo = bar;" where foo and bar are the
same kind of struct, then we don't need to set all the members of foo to
user data, we can use the fake assignments instead.  The bug here is that
maybe we already set parts of bar to non-user data so setting all of foo
to user data causes false positives.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>