user_data2: always set *foo to user data
Say you have:

	if (copy_from_user(&foo, src, sizeof(foo))) ...

Before if "foo" was not a struct we would set *foo to user data, but if
it was a struct then we would set all the members foo->bar, foo->baz to
user data.  Now we set all the members but we also set *foo as well.  The
reason is that if we cast foo so we have different member names then we
need to assume those members are user data.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>