memcpy_overflow: silence a bunch of false positives
There are several strategies here that sort of overlap and it's all not
perfect but it generally silences false positives so it's a step forward.
It's a work in progress.

The main thing here is if we copy several members of a struct in one
memcpy() that is normally intentional.

If we are copying data from a member of a union then sometimes we copy
more than the member.  This is normally a false positive.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>