modification_hooks: introduce  add_modification_hook_late()
Normally what happens is we have:

	p = some_function_returns_ERR_PTR();
	if (IS_ERR(p)) {

So we want the modification hooks to run before we set p to an ERR_PTR.
But I have some new code coming where we do:

	if (foo + bar > baz)
		foo = baz - bar;

Smatch comparison stores the comparison that foo > baz - bar but the
problem is that when we go to use that in smatch extra the modification
hook clears foo before we can use the comparison.  This patch introduces a
new modification hook that runs after smatch_extra.c.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>