comparison: make modifications clear the comparisons later
This uses the new add_modification_hook_late() and also move the
match_assign hook later so that they happen after smatch extra.

The issue is that we have:

	if (foo + bar > size)
		foo = size - bar;

We know that foo is more than size - bar and we want to use it in Smatch
extra.  But the information was cleared right before we could use it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>