extra: handle if (offset + bytes > size) comparisons better
The code we are trying to parse looks like:

	if (offset + bytes > size)
		bytes = size - offset;

It's simple algebra and there are several ways we could handle it.  What I
have done is that I've used smatch_comparison to say that we know
"size - offset" is less than bytes so if we know what bytes started as we
can go from there.

I might as well later update how extra handles conditions to do more
algebra than it currently does.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>