modification_hooks: fix fall out from recent changes
The recent change of splitting the modifications up into late and early
caused a bunch of false positives.  I wrote this to fix it.  Now that I'm
committing the code, I'm not totally sure it's the right thing...  It works
but maybe I should re-write it cleaner and change the call sites etc.

Future work I suppose.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>