memcpy_overflow: fall back to type size to calculate bytes to end of struct
I probably should change get_array_size_bytes_max() here instead of this...

The situation is, say you have:

	struct foo *p = buf;

	memcpy(&p->member, src, size);

But smatch is saying that we don't know the actual size of p because we
don't know the size of buf in this example.  To work around it, say that
if we don't know the size from buf, then just use sizeof(struct foo).

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>