function_hooks: don't process condition functions twice
When we had a function that was used as a condition then it was first
processed as a normal function call then as part of a condition.  This is
a waste of time, but the real reason for this commit is because I want to
use the implications of the function call and the first processing was
overwriting them.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>