indenting: ignore pre and post statements
Hm...  This looks serious, but I think it's not.  What happened was that
__cur_stmt does not always point to the *current* statement.  What it does
is that it was designed to basically look at each line of code.  So if you
have:

	for (i = 0; i < 10; i++) {

Then it count's that as one statement because it's basically a for
statement.  But if you want to be technical there are is a pre and post
statement in there.  And I wanted to count those so I could always find
the exact current statement.

But I got distracted part way through and there was a lot of code which
relied on the old behavior...

But I want to commit this anyway, because I'm carrying around too big of
a diff.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>