implied: use implications from function conditions
Say we have a function:

int is_ten(int x)
{
	if (x == 10)
		return 1;
	return 0;
}

Then the condition "if (x == 10) " and "if (is_ten(x)) " are sort of the
same.  But in Smatch we could only calculate the implications of the
direct compare and not the function call.

Comparisons like this are translated into PARAM_LIMIT entries in the
database.  The PARAM_LIMIT entries are processed very early after
PARAM_CLEARED entries.  I thought about moving it to be processed first
but I don't think it matters.  (I'm not 100% positive).

So basically we take a PARAM_LIMIT for a side, assume it's true then see
what that implies.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>