extra: fix values_fit_type()
This function is used when we have code like:

	x = y;

We want to return true if all the possible values for y can be stored in
x.  For example, if y is an int in range 0-255 and x unsigned char then
they fit.  But there was a bug in handling type promotion so it was saying
that negative numbers could fit into unsigned int.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>