validation/sm_implied10.c: update the test
These days if we use an variable as an offset into an array, we afterwards
assume that it is a valid index.  In theory, it's the same as assuming that
if we dereferenced a variable it means it's non-NULL...  But actually I'm
sort on the fence about this.  But that's how the code is written so lets
update the validation tests to match.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>