implied: fix a bug handling PARAM_LIMIT implications
The first thing is that in Smatch function hooks I added a nullify/unnulify
path.  That's because I'm trying to reset the cur_stree back to the
original but it only cleared states which were in the original stree and
not newly introduced states.

The other problem is that set_state and __set_sm() sets the cur_stree and
the true and false strees.  Most of the time it's fine because we immediately
overwrite it with the correct values.  But if we only want to set a true
state and not a false state then it doesn't work.  So I have introduced a
new __set_sm_cur_stree() which only sets the stree.

I probably need to check all the __set_sm() callers.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>